<details>
<summary>After this PR: 7.818833ms</summary>

```sh
$ ./four-keys releases --repository https://github.com/go-git/go-git --since 2020-01-01 --until 2020-12-31 --debug
[Debug] In debug mode
[Debug] StartTimer: Open Repository
[Debug] Stop_Timer: Open Repository      1.378716167s
[Debug] StartTimer: QueryReleases
[Debug] StartTimer: QueryTags
[Debug] Stop_Timer: QueryTags    20.292µs
[Debug] Tags count: 60
[Debug] Sources count: 60
[Debug] [Debug] source[ 14 ]( v4.8.1 ) is skipped for outof time range
[Debug] source[ 48 ]( v3.0.3 ) is skipped for outof time range
[Debug] source[ 2 ]( v5.4.0 ) is skipped for outof time range
[Debug] source[ 55 ]( v2.1.2 ) is skipped for outof time range
[Debug] StartTimer: source[4](v5.2.0)GetEachReleases
[Debug] source[ 56 ]( v2.1.1 ) is skipped for outof time range
[Debug] [Debug] source[ 36 ]( v4.0.0-rc8 ) is skipped for outof time range
source[ 42 ]( v4.0.0-rc2 ) is skipped for outof time range
[Debug] [Debug] [Debug] source[ 43 ]( v4.0.0-rc1 ) is skipped for outof time range
[Debug] source[ 15 ]( v4.8.0 ) is skipped for outof time range
[Debug] [Debug] source[ 44 ]( v3.2.0 ) is skipped for outof time range
[Debug] source[ 16 ]( v4.7.1 ) is skipped for outof time range
[Debug] source[ 38 ]( v4.0.0-rc6 ) is skipped for outof time range
[Debug] [Debug] source[ 53 ]( v2.2.0 ) is skipped for outof time range
[Debug] source[ 18 ]( v4.6.0 ) is skipped for outof time range
[Debug] source[ 40 ]( v4.0.0-rc4 ) is skipped for outof time range
[Debug] source[ 49 ]( v3.0.2 ) is skipped for outof time range
[Debug] [Debug] [Debug] source[ 19 ]( v4.5.0 ) is skipped for outof time range
[Debug] source[ 45 ]( v3.1.1 ) is skipped for outof time range
[Debug] source[ 30 ]( v4.0.0-rc14 ) is skipped for outof time range
[Debug] source[ 51 ]( v3.0.0 ) is skipped for outof time range
[Debug] [Debug] source[ 24 ]( v4.2.1 ) is skipped for outof time range
[Debug] source[ 52 ]( v2.2.1 ) is skipped for outof time range
[Debug] [Debug] [Debug] source[ 33 ]( v4.0.0-rc11 ) is skipped for outof time range
[Debug] source[ 39 ]( v4.0.0-rc5 ) is skipped for outof time range
[Debug] source[ 8 ]( v4.13.0 ) is skipped for outof time range
source[ 46 ]( v3.1.0 ) is skipped for outof time range
source[ 59 ]( v1.0.0 ) is skipped for outof time range
[Debug] source[ 7 ]( v4.13.1 ) is skipped for outof time range
source[ 23 ]( v4.3.0 ) is skipped for outof time range
[Debug] source[ 31 ]( v4.0.0-rc13 ) is skipped for outof time range
source[ 25 ]( v4.2.0 ) is skipped for outof time range
[Debug] source[ 20 ]( v4.4.1 ) is skipped for outof time range
source[ 37 ]( v4.0.0-rc7 ) is skipped for outof time range
[Debug] source[ 3 ]( v5.3.0 ) is skipped for outof time range
source[ 41 ]( v4.0.0-rc3 ) is skipped for outof time range
[Debug] source[ 32 ]( v4.0.0-rc12 ) is skipped for outof time range
[Debug] source[ 27 ]( v4.1.0 ) is skipped for outof time range
source[ 10 ]( v4.11.0 ) is skipped for outof time range
[Debug] source[ 29 ]( v4.0.0-rc15 ) is skipped for outof time range
[Debug] StartTimer: source[6](v5.0.0)GetEachReleases
[Debug] source[ 26 ]( v4.1.1 ) is skipped for outof time range
[Debug] source[ 1 ]( v5.4.1 ) is skipped for outof time range
[Debug] [Debug] StartTimer: source[5](v5.1.0)GetEachReleases
[Debug] source[ 17 ]( v4.7.0 ) is skipped for outof time range
source[ 11 ]( v4.10.0 ) is skipped for outof time range
[Debug] source[ 50 ]( v3.0.1 ) is skipped for outof time range
[Debug] source[ 57 ]( v2.1.0 ) is skipped for outof time range
[Debug] source[ 13 ]( v4.9.0 ) is skipped for outof time range
[Debug] [Debug] source[ 35 ]( v4.0.0-rc9 ) is skipped for outof time range
source[ 0 ]( v5.4.2 ) is skipped for outof time range
[Debug] source[ 34 ]( v4.0.0-rc10 ) is skipped for outof time range
[Debug] source[ 47 ]( v3.0.4 ) is skipped for outof time range
[Debug] source[ 22 ]( v4.3.1 ) is skipped for outof time range
[Debug] source[ 58 ]( v2.0.0 ) is skipped for outof time range
[Debug] source[ 21 ]( v4.4.0 ) is skipped for outof time range
source[ 54 ]( v2.1.3 ) is skipped for outof time range
source[ 9 ]( v4.12.0 ) is skipped for outof time range
source[ 12 ]( v4.9.1 ) is skipped for outof time range
[Debug] source[ 28 ]( v4.0.0 ) is skipped for outof time range
[Debug] Stop_Timer: source[6](v5.0.0)GetEachReleases     6.542125ms
[Debug] Stop_Timer: source[4](v5.2.0)GetEachReleases     7.162458ms
[Debug] Stop_Timer: source[5](v5.1.0)GetEachReleases     7.295792ms
[Debug] StartTimer: release[0](v5.2.0)GetReleaseMetrics
[Debug] Stop_Timer: release[0](v5.2.0)GetReleaseMetrics          1.875µs
[Debug] StartTimer: release[1](v5.1.0)GetReleaseMetrics
[Debug] Stop_Timer: release[1](v5.1.0)GetReleaseMetrics          1.209µs
[Debug] StartTimer: release[2](v5.0.0)GetReleaseMetrics
[Debug] Stop_Timer: release[2](v5.0.0)GetReleaseMetrics          1.167µs
[Debug] Stop_Timer: QueryReleases        7.818833ms
{"option":{"since":"2020-01-01T00:00:00Z","until":"2020-12-31T23:59:59Z"},"releases":[{"tag":"v5.2.0","date":"2020-10-09T11:49:30+02:00","leadTimeForChanges":{"value":130.77916666666667,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}},{"tag":"v5.1.0","date":"2020-05-24T19:25:08+02:00","leadTimeForChanges":{"value":69.86515046296296,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}},{"tag":"v5.0.0","date":"2020-03-15T21:18:32+01:00","leadTimeForChanges":{"value":224.73468749999998,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}}]}⏎                
```

</details>

<details>
<summary>Before this PR: 17.281292ms</summary>

```sh
$ ./four-keys releases --repository https://github.com/go-git/go-git --since 2020-01-01 --until 2020-12-31 --debug
[Debug] In debug mode
[Debug] StartTimer: Open Repository
[Debug] Stop_Timer: Open Repository      1.087360417s
[Debug] StartTimer: QueryReleases
[Debug] StartTimer: QueryTags
[Debug] Stop_Timer: QueryTags    11.541µs
[Debug] Tags count: 60
[Debug] Sources count: 60
[Debug] source[ 0 ]( v5.4.2 ) is skipped for outof time range
[Debug] source[ 1 ]( v5.4.1 ) is skipped for outof time range
[Debug] source[ 2 ]( v5.4.0 ) is skipped for outof time range
[Debug] source[ 3 ]( v5.3.0 ) is skipped for outof time range
[Debug] StartTimer: source[4](v5.2.0)GetReleaseMetrics
[Debug] Stop_Timer: source[4](v5.2.0)GetReleaseMetrics   6.174625ms
[Debug] StartTimer: source[5](v5.1.0)GetReleaseMetrics
[Debug] Stop_Timer: source[5](v5.1.0)GetReleaseMetrics   4.997292ms
[Debug] StartTimer: source[6](v5.0.0)GetReleaseMetrics
[Debug] Stop_Timer: source[6](v5.0.0)GetReleaseMetrics   5.239667ms
[Debug] source[ 7 ]( v4.13.1 ) is skipped for outof time range
[Debug] source[ 8 ]( v4.13.0 ) is skipped for outof time range
[Debug] source[ 9 ]( v4.12.0 ) is skipped for outof time range
[Debug] source[ 10 ]( v4.11.0 ) is skipped for outof time range
[Debug] source[ 11 ]( v4.10.0 ) is skipped for outof time range
[Debug] source[ 12 ]( v4.9.1 ) is skipped for outof time range
[Debug] source[ 13 ]( v4.9.0 ) is skipped for outof time range
[Debug] source[ 14 ]( v4.8.1 ) is skipped for outof time range
[Debug] source[ 15 ]( v4.8.0 ) is skipped for outof time range
[Debug] source[ 16 ]( v4.7.1 ) is skipped for outof time range
[Debug] source[ 17 ]( v4.7.0 ) is skipped for outof time range
[Debug] source[ 18 ]( v4.6.0 ) is skipped for outof time range
[Debug] source[ 19 ]( v4.5.0 ) is skipped for outof time range
[Debug] source[ 20 ]( v4.4.1 ) is skipped for outof time range
[Debug] source[ 21 ]( v4.4.0 ) is skipped for outof time range
[Debug] source[ 22 ]( v4.3.1 ) is skipped for outof time range
[Debug] source[ 23 ]( v4.3.0 ) is skipped for outof time range
[Debug] source[ 24 ]( v4.2.1 ) is skipped for outof time range
[Debug] source[ 25 ]( v4.2.0 ) is skipped for outof time range
[Debug] source[ 26 ]( v4.1.1 ) is skipped for outof time range
[Debug] source[ 27 ]( v4.1.0 ) is skipped for outof time range
[Debug] source[ 28 ]( v4.0.0 ) is skipped for outof time range
[Debug] source[ 29 ]( v4.0.0-rc15 ) is skipped for outof time range
[Debug] source[ 30 ]( v4.0.0-rc14 ) is skipped for outof time range
[Debug] source[ 31 ]( v4.0.0-rc13 ) is skipped for outof time range
[Debug] source[ 32 ]( v4.0.0-rc12 ) is skipped for outof time range
[Debug] source[ 33 ]( v4.0.0-rc11 ) is skipped for outof time range
[Debug] source[ 34 ]( v4.0.0-rc10 ) is skipped for outof time range
[Debug] source[ 35 ]( v4.0.0-rc9 ) is skipped for outof time range
[Debug] source[ 36 ]( v4.0.0-rc8 ) is skipped for outof time range
[Debug] source[ 37 ]( v4.0.0-rc7 ) is skipped for outof time range
[Debug] source[ 38 ]( v4.0.0-rc6 ) is skipped for outof time range
[Debug] source[ 39 ]( v4.0.0-rc5 ) is skipped for outof time range
[Debug] source[ 40 ]( v4.0.0-rc4 ) is skipped for outof time range
[Debug] source[ 41 ]( v4.0.0-rc3 ) is skipped for outof time range
[Debug] source[ 42 ]( v4.0.0-rc2 ) is skipped for outof time range
[Debug] source[ 43 ]( v4.0.0-rc1 ) is skipped for outof time range
[Debug] source[ 44 ]( v3.2.0 ) is skipped for outof time range
[Debug] source[ 45 ]( v3.1.1 ) is skipped for outof time range
[Debug] source[ 46 ]( v3.1.0 ) is skipped for outof time range
[Debug] source[ 47 ]( v3.0.4 ) is skipped for outof time range
[Debug] source[ 48 ]( v3.0.3 ) is skipped for outof time range
[Debug] source[ 49 ]( v3.0.2 ) is skipped for outof time range
[Debug] source[ 50 ]( v3.0.1 ) is skipped for outof time range
[Debug] source[ 51 ]( v3.0.0 ) is skipped for outof time range
[Debug] source[ 52 ]( v2.2.1 ) is skipped for outof time range
[Debug] source[ 53 ]( v2.2.0 ) is skipped for outof time range
[Debug] source[ 54 ]( v2.1.3 ) is skipped for outof time range
[Debug] source[ 55 ]( v2.1.2 ) is skipped for outof time range
[Debug] source[ 56 ]( v2.1.1 ) is skipped for outof time range
[Debug] source[ 57 ]( v2.1.0 ) is skipped for outof time range
[Debug] source[ 58 ]( v2.0.0 ) is skipped for outof time range
[Debug] source[ 59 ]( v1.0.0 ) is skipped for outof time range
[Debug] Stop_Timer: QueryReleases        17.281292ms
{"option":{"since":"2020-01-01T00:00:00Z","until":"2020-12-31T23:59:59Z"},"releases":[{"tag":"v5.2.0","date":"2020-10-09T11:49:30+02:00","leadTimeForChanges":{"value":130.77916666666667,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}},{"tag":"v5.1.0","date":"2020-05-24T19:25:08+02:00","leadTimeForChanges":{"value":69.86515046296296,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}},{"tag":"v5.0.0","date":"2020-03-15T21:18:32+01:00","leadTimeForChanges":{"value":224.73468749999998,"unit":"day"},"result":{"isSuccess":true,"timeToRestore":null}}]}⏎    
```

</details>